Completion of Customer, Progress on Manager
The Customer JS communicates with the Database, updating if a customer makes a purchase (While informing them of the price of purchase) or informing the customer of not enough items in stock.

The Manager script is currently in progress with view all products being the only active feature. The steps for low inventory is currently being worked on.